refactor(runtime): move mid-turn compaction and overflow recovery into AiSdkCompaction
Move the mid-turn capacity family (buildMidTurnCapacityCompactState,
buildMidTurnCapacityCompactPrepareStep, computeMidTurnCompactionReplacement,
recoverFromOverflowError, buildMidTurnFinalRequestVerdict), the
MidTurnCapacityCompactState class, and the mid-turn helpers from AiSdkBackend
into the AiSdkCompaction collaborator. hasBlockingReplayDiagnostics moves and is
exported back to AiSdkBackend (still used by buildPriorMessages).

The replay/tail host seam (materializeRuntimeReplayPlan, canReplayProviderNative,
appendTurnTailPrompt) stays on AiSdkBackend and is injected as three callbacks;
call sites keep their this.* names. Turn-scoped abort state is threaded per call:
abortSignal into the capacity prepareStep, computeMidTurnCompactionReplacement,
and overflow recovery; abortController into the final-request verdict (it calls
.abort()). MidTurnCapacityCompactState remains a directly-written shared object
owned by both send() and the orchestrator methods.

Behavior-neutral pure move: 2073 pass / 0 fail / 7 skipped.

Part of #1084
