+++ packages/runtime/src/ai-sdk-compaction.ts
@@ -13,6 +13,7 @@ import type { RuntimeEvent } from '@maka/core/runtime-event';
 import type {
   BackendCompactHistoryInput,
   BackendCompactHistoryResult,
+  BackendSendInput,
 } from '@maka/core/backend-types';
 import type { ContextBudgetDiagnostic, LlmCallRecord } from '@maka/core/usage-stats/types';
 
@@ -35,7 +36,10 @@ import {
   type SynthesisSourceRef,
   type ToolResultArchiveRef,
 } from './context-budget.js';
-import { evaluateHistoryCompactCheckpointReplay } from './history-compact.js';
+import {
+  evaluateHistoryCompactCheckpointReplay,
+  isHistoryCompactContentEvent,
+} from './history-compact.js';
 import { HistoryCompactSummarizerError } from './history-compact-error.js';
 import {
   buildHistoryCompactCheckpoint,
@@ -45,12 +49,13 @@ import {
 
 import { createHash } from 'node:crypto';
 import type { ModelMessage } from 'ai';
-import type {
-  ModelAdapter,
-  NormalizedAiSdkUsage,
-  PrepareStepFunctionLike,
-  PrepareStepLike,
-  PrepareStepResultLike,
+import {
+  normalizeAiSdkUsage,
+  type ModelAdapter,
+  type NormalizedAiSdkUsage,
+  type PrepareStepFunctionLike,
+  type PrepareStepLike,
+  type PrepareStepResultLike,
 } from './model-adapter.js';
 import {
   activeToolResultLineageIdentity,
@@ -70,6 +75,22 @@ import {
 } from './semantic-compact.js';
 import { collectStaleToolResultArchiveCandidates } from './tool-result-archive.js';
 
+import type { ContextBudgetExhaustedDetail, SessionEvent } from '@maka/core/events';
+import type { AsyncEventQueue } from './async-queue.js';
+import type { MakaTool } from './tool-runtime.js';
+import {
+  buildRuntimeEventModelReplayPlan,
+  collectToolActivityTurnIds,
+  type RuntimeEventModelReplayPlan,
+} from './model-history.js';
+import { toolSchemaCharsForDiagnostics } from './request-shape.js';
+import {
+  estimateNextRequestTokens,
+  exceedsHighWater,
+  planMidTurnCapacityCompaction,
+} from './mid-turn-capacity-compact.js';
+import { resolveSelectedModelContextWindow } from './context-budget-policy.js';
+
 /** Constructor dependencies for AiSdkCompaction. Grows as more families move in. */
 export interface AiSdkCompactionDeps {
   input: AiSdkBackendInput;
@@ -77,6 +98,12 @@ export interface AiSdkCompactionDeps {
   now: () => number;
   modelAdapter: ModelAdapter;
   computeCostUsd: (usage: NormalizedAiSdkUsage) => number | undefined;
+  materializeRuntimeReplayPlan: (plan: RuntimeEventModelReplayPlan) => Promise<ModelMessage[]>;
+  canReplayProviderNative: (plan: RuntimeEventModelReplayPlan) => boolean;
+  appendTurnTailPrompt: (
+    content: ModelMessage['content'],
+    turnTailPrompt?: string,
+  ) => ModelMessage['content'];
 }
 
 export class AiSdkCompaction {
@@ -85,6 +112,14 @@ export class AiSdkCompaction {
   private readonly now: () => number;
   private readonly modelAdapter: ModelAdapter;
   private readonly computeCostUsd: (usage: NormalizedAiSdkUsage) => number | undefined;
+  private readonly materializeRuntimeReplayPlan: (
+    plan: RuntimeEventModelReplayPlan,
+  ) => Promise<ModelMessage[]>;
+  private readonly canReplayProviderNative: (plan: RuntimeEventModelReplayPlan) => boolean;
+  private readonly appendTurnTailPrompt: (
+    content: ModelMessage['content'],
+    turnTailPrompt?: string,
+  ) => ModelMessage['content'];
   private historyCompactAbortController: AbortController | null = null;
 
   constructor(deps: AiSdkCompactionDeps) {
@@ -93,6 +128,9 @@ export class AiSdkCompaction {
     this.now = deps.now;
     this.modelAdapter = deps.modelAdapter;
     this.computeCostUsd = deps.computeCostUsd;
+    this.materializeRuntimeReplayPlan = deps.materializeRuntimeReplayPlan;
+    this.canReplayProviderNative = deps.canReplayProviderNative;
+    this.appendTurnTailPrompt = deps.appendTurnTailPrompt;
   }
 
   /** Abort an in-flight manual history compaction (called by AiSdkBackend.stop). */
@@ -1169,6 +1207,715 @@ export class AiSdkCompaction {
       // perturb provider request projection or tool-loop progress.
     }
   }
+
+  public buildMidTurnCapacityCompactState(
+    input: BackendSendInput,
+  ): MidTurnCapacityCompactState | undefined {
+    const policy = this.input.contextBudget;
+    if (
+      policy?.historyCompact?.enabled !== true ||
+      policy.historyCompact.midTurn?.enabled !== true
+    ) {
+      return undefined;
+    }
+    if (
+      !this.input.summarizeHistoryCompact ||
+      !this.input.recordHistoryCompactCheckpoint ||
+      !this.input.loadTurnRuntimeEvents
+    ) {
+      return undefined;
+    }
+    const headAnchor = input.headAnchorRuntimeEvent;
+    if (
+      !headAnchor ||
+      headAnchor.sessionId !== this.sessionId ||
+      headAnchor.turnId !== input.turnId ||
+      headAnchor.role !== 'user' ||
+      headAnchor.author !== 'user' ||
+      !isHistoryCompactContentEvent(headAnchor)
+    ) {
+      return undefined;
+    }
+    const contextWindow = resolveSelectedModelContextWindow(
+      this.input.connection,
+      this.input.modelId,
+    );
+    if (contextWindow === undefined) return undefined;
+    const priorContentEvents = (input.runtimeContext ?? [])
+      .filter((event) => event.turnId !== input.turnId)
+      .filter(isHistoryCompactContentEvent);
+    return new MidTurnCapacityCompactState(headAnchor, priorContentEvents, contextWindow);
+  }
+
+  /**
+   * prepareStep SHAPING hook for the mid-turn capacity invariant: between
+   * steps of one turn, estimate the next provider request (last step's real
+   * usage + a signed char/4 payload delta, tool schemas included) against
+   * `contextWindow - reserve`; over the high-water, fold a safe completed
+   * prefix into a durable mid_turn checkpoint and continue the same turn on
+   * `[compact block, verbatim head anchor, preserved tail]`.
+   *
+   * This hook never terminates the turn: every failure fails open with a
+   * diagnostic and records itself for the final-request estimate owner, which
+   * re-measures the payload after ALL shaping (including active tool-result
+   * pruning, which runs later and can still rescue the step) and issues the
+   * context_budget_exhausted verdict only when the request that would really
+   * go out exceeds the window. The trigger threshold here is deliberately
+   * approximate — a missed or spurious trigger is recoverable; the verdict is
+   * not, so it does not live here.
+   */
+  public buildMidTurnCapacityCompactPrepareStep(
+    turnId: string,
+    state: MidTurnCapacityCompactState | undefined,
+    queue: AsyncEventQueue<SessionEvent>,
+    providerTools: readonly MakaTool[],
+    fallbackActiveTools: () => readonly string[],
+    turnTailPrompt: string | undefined,
+    systemPromptChars: number,
+    onDiagnosticPatch: (patch: Partial<ContextBudgetDiagnostic>) => void,
+    abortSignal?: AbortSignal,
+  ): PrepareStepFunctionLike | undefined {
+    if (!state) return undefined;
+    const policy = this.input.contextBudget!;
+    const compactPolicy = policy.historyCompact!;
+    const midTurn = compactPolicy.midTurn!;
+    const charsPerToken = policy.charsPerToken ?? 4;
+    const reserveTokens = midTurn.reserveTokens ?? 16_384;
+    let acceptedProjection: ActiveFullCompactPrepareStepProjection | undefined;
+
+    return async (options) => {
+      const incomingMessages = options.messages;
+      const projectedMessages = projectAcceptedActiveFullCompactMessages(
+        incomingMessages,
+        acceptedProjection,
+      );
+      const keepProjection = (): PrepareStepResultLike | undefined =>
+        projectedMessages ? { messages: projectedMessages } : undefined;
+      // Step 0 is shaped by the pre_turn path; the mid-turn trigger only runs
+      // between steps, once completed-step usage and events exist.
+      if (options.stepNumber < 1 || state.exhaustedDetail) return keepProjection();
+
+      // Real usage for the last finished step, read synchronously from the
+      // SDK's own step results (the same numbers the finish-step chunk
+      // carries) — no coupling to how far the stream consumer has advanced.
+      // Baseline = the last request's INPUT tokens only (see the state field
+      // doc: the payload delta already carries the step's output). The
+      // adapter fails closed on missing token counts (undefined, #972), and a
+      // provider can still report a zero input outright — either way a
+      // non-positive input count is unusable for estimation, so clear the
+      // baseline and let the estimate fall back to the whole-payload cold
+      // start instead of "0 + delta".
+      //
+      // The usage anchor is only meaningful PAIRED with the payload baseline
+      // of the request it was reported for (`lastRequestPayloadChars`). A
+      // successful overflow recovery restructures the request and resets that
+      // baseline to undefined: the send-global steps view still carries the
+      // dead attempt's last usage, but anchoring on it against the rejected
+      // request's chars would under-estimate the retry by the whole previous
+      // step growth — so a missing baseline forces the whole-payload cold
+      // start, exactly like a missing usage sample.
+      const lastStepInputTokens = normalizeAiSdkUsage(options.steps.at(-1)?.usage)?.inputTokens;
+      state.lastRequestInputTokens =
+        state.lastRequestPayloadChars !== undefined &&
+        lastStepInputTokens !== undefined &&
+        Number.isFinite(lastStepInputTokens) &&
+        lastStepInputTokens > 0
+          ? lastStepInputTokens
+          : undefined;
+
+      // A skipped trigger is never silent: every failure-driven skip records a
+      // failedOpen decision. Recorder counters are attached ONLY on the tiers
+      // where the recorder was actually invoked — the diagnostics must never
+      // claim a write that did not happen.
+      const failOpen = (
+        failOpenReason: string,
+        recorderCounters: Partial<ContextBudgetDiagnostic> = {},
+      ): PrepareStepResultLike | undefined => {
+        onDiagnosticPatch({
+          historyCompactEnabled: true,
+          historyCompactMode: 'read_write',
+          ...recorderCounters,
+          ...compactionDecisionDiagnosticPatch({
+            stage: 'activeStep',
+            sourceKind: 'runtimeEvents',
+            decision: 'failedOpen',
+            phase: 'mid_turn',
+            boundaryKind: 'historyCompact',
+            reason: 'context_limit',
+            failOpenReason,
+            skippedReasonCounts: { [failOpenReason]: 1 },
+          }),
+        });
+        return keepProjection();
+      };
+      // A shaping failure additionally records itself for the final-request
+      // estimate owner: when the final payload is still over the window, the
+      // owner turns this step's failure into the terminal detail instead of
+      // re-entering a shaper that already attempted and failed.
+      const shapeFailure = (
+        detail: ContextBudgetExhaustedDetail,
+        diagnosticReason: string,
+        recorderCounters: Partial<ContextBudgetDiagnostic> = {},
+      ): PrepareStepResultLike | undefined => {
+        state.lastShapeFailure = { stepNumber: options.stepNumber, detail, diagnosticReason };
+        return failOpen(diagnosticReason, recorderCounters);
+      };
+
+      // Trigger estimate: the last request's input tokens plus a SIGNED char/4 delta of
+      // this step's payload (system prompt + projected messages + active tool
+      // schemas) against the previous request's measured payload. Measured synchronously from
+      // the SDK's own projection — no ledger dependency — so a same-turn
+      // `load_tools` schema expansion or a large tool result both count. This
+      // position measures BEFORE later shapers (prune) run, so it can
+      // over-trigger; that is the recoverable direction, and the verdict owner
+      // re-measures the post-shaping payload.
+      const measuredMessages = projectedMessages ?? incomingMessages;
+      const activeToolsForStep = options.activeTools ?? fallbackActiveTools();
+      const payloadChars = midTurnRequestPayloadChars(
+        measuredMessages,
+        providerTools,
+        activeToolsForStep,
+        systemPromptChars,
+      );
+      const forcedEstimate = state.forcedTriggerEstimate;
+      state.forcedTriggerEstimate = undefined;
+      const estimate =
+        forcedEstimate ??
+        estimateNextRequestTokens({
+          ...(state.lastRequestInputTokens !== undefined
+            ? { priorUsageTokens: state.lastRequestInputTokens }
+            : {}),
+          appendedChars: payloadChars - (state.lastRequestPayloadChars ?? payloadChars),
+          charsPerToken,
+          coldStartChars: payloadChars,
+        });
+      if (
+        forcedEstimate === undefined &&
+        !exceedsHighWater(estimate, state.contextWindow, reserveTokens)
+      ) {
+        return keepProjection();
+      }
+
+      // Fold a safe completed prefix of the durable turn ledger into a
+      // replacement projection (validate → persist), shared with the reactive
+      // overflow path. This hook maps the outcome to the prepareStep contract:
+      // keep the raw projection on skip/fail, apply the fold on success.
+      const outcome = await this.computeMidTurnCompactionReplacement({
+        turnId,
+        state,
+        queue,
+        minFlushedSteps: options.stepNumber,
+        estimatedNextRequestTokens: estimate,
+        referencePayloadChars: payloadChars,
+        providerTools,
+        activeToolsForStep,
+        systemPromptChars,
+        turnTailPrompt,
+        abortSignal,
+      });
+      if (outcome.decision === 'skip') return keepProjection();
+      if (outcome.decision === 'fail') {
+        return shapeFailure(outcome.detail, outcome.diagnosticReason, outcome.recorderCounters);
+      }
+      acceptedProjection = {
+        sourceSignatures: incomingMessages.map(modelMessageSignature),
+        sourceSignatureMode: 'exact',
+        projectedMessages: outcome.replacementMessages,
+      };
+      state.replacedStepNumber = options.stepNumber;
+      onDiagnosticPatch(
+        buildMidTurnReplacedDiagnosticPatch({
+          checkpoint: outcome.checkpoint,
+          estimatedTokensBefore: outcome.estimatedTokensBefore,
+          estimatedTokensAfter: outcome.estimatedTokensAfter,
+          reason: 'context_limit',
+        }),
+      );
+      return { messages: outcome.replacementMessages };
+    };
+  }
+
+  /**
+   * Fold a safe completed prefix of the durable turn ledger into a persisted
+   * mid_turn checkpoint and its `[block, verbatim anchor, tail]` replacement
+   * messages — the compaction core shared by the proactive prepareStep hook
+   * (issue #882 PR 1) and the reactive overflow recovery (PR 2). It waits for
+   * the seq-ack durability boundary, reads the ledger, plans the fold, then
+   * validates (materializable ∧ smaller than the reference request ∧
+   * replay-admissible) and persists BEFORE returning the replacement, so a
+   * recovery re-projection never re-injects a covered raw span. It only shapes:
+   * the pass/terminate verdict and the diagnostic emission are the caller's.
+   */
+  public async computeMidTurnCompactionReplacement(input: {
+    turnId: string;
+    state: MidTurnCapacityCompactState;
+    queue: AsyncEventQueue<SessionEvent>;
+    minFlushedSteps: number;
+    estimatedNextRequestTokens: number;
+    referencePayloadChars: number;
+    providerTools: readonly MakaTool[];
+    activeToolsForStep: readonly string[];
+    systemPromptChars: number;
+    turnTailPrompt: string | undefined;
+    abortSignal?: AbortSignal;
+  }): Promise<MidTurnCompactionOutcome> {
+    const {
+      turnId,
+      state,
+      queue,
+      providerTools,
+      activeToolsForStep,
+      systemPromptChars,
+      turnTailPrompt,
+      abortSignal,
+    } = input;
+    const summarizer = this.input.summarizeHistoryCompact!;
+    const recorder = this.input.recordHistoryCompactCheckpoint!;
+    const loadTurnRuntimeEvents = this.input.loadTurnRuntimeEvents!;
+    const policy = this.input.contextBudget!;
+    const compactPolicy = policy.historyCompact!;
+    const midTurn = compactPolicy.midTurn!;
+    const charsPerToken = policy.charsPerToken ?? 4;
+    const reserveTokens = midTurn.reserveTokens ?? 16_384;
+
+    // Coverage pool = the durable run ledger, read through the injected
+    // seam. Covered events are persisted by construction (no crash window
+    // between checkpoint and source), and their bytes are exactly what a
+    // recovery re-projection replays.
+    //
+    // Seq-ack durability boundary. The replacement projection REPLACES the
+    // whole message list, so any completed-step content event missing from
+    // the durable pool is silently dropped from the next request — a
+    // lagging ledger here is content loss (e.g. a step's already-emitted
+    // assistant text), not a conservative under-count. No event-kind
+    // predicate can close that: the wait counts the event stream itself.
+    //  1. The pump has flushed every finish-step boundary the SDK reports
+    //     completed (state.flushedSteps), so ALL of the completed steps'
+    //     session events — tool pairs AND thinking/text completions — are
+    //     enqueued with producer-stamped sequence numbers.
+    //  2. The consumer has fully processed everything enqueued
+    //     (consumedCount >= pushedCount). The consumer's pull is the ack
+    //     (see drain()): it fires after processing, not after persisting,
+    //     so deliberately-unpersisted events (non-terminal errors,
+    //     partials) can never deadlock the wait.
+    // After both, ONE durable read (which itself re-awaits the run's
+    // serialized write queue) sees every event the projection may carry.
+    // Exits: the boundary, an abort, a detached consumer, or a read failure.
+    for (;;) {
+      if (abortSignal?.aborted) {
+        return {
+          decision: 'fail',
+          detail: 'no_safe_completed_span',
+          diagnosticReason: 'ledger_wait_aborted',
+        };
+      }
+      if (queue.consumerDetached) {
+        return {
+          decision: 'fail',
+          detail: 'no_safe_completed_span',
+          diagnosticReason: 'ledger_wait_aborted',
+        };
+      }
+      if (state.flushedSteps >= input.minFlushedSteps && queue.consumedCount >= queue.pushedCount)
+        break;
+      await waitForQueueProgressOrAbort(queue, abortSignal);
+    }
+    let turnLedger: RuntimeEvent[];
+    try {
+      turnLedger = await loadTurnRuntimeEvents(turnId);
+    } catch {
+      return {
+        decision: 'fail',
+        detail: 'no_safe_completed_span',
+        diagnosticReason: 'ledger_read_failed',
+      };
+    }
+    const currentTurnEvents = turnLedger
+      .filter((event) => event.turnId === turnId)
+      .filter(isHistoryCompactContentEvent);
+    // The head anchor is persisted before backend.send() is invoked, so
+    // its absence is a wiring error, not replication lag — fail open now.
+    if (!currentTurnEvents.some((event) => event.id === state.headAnchor.id)) {
+      return {
+        decision: 'fail',
+        detail: 'no_safe_completed_span',
+        diagnosticReason: 'head_anchor_not_durable',
+      };
+    }
+    const orderedEvents = [...state.priorContentEvents, ...currentTurnEvents];
+
+    const plan = await planMidTurnCapacityCompaction({
+      sessionId: this.sessionId,
+      orderedEvents,
+      headAnchor: { runtimeEventId: state.headAnchor.id, turnId },
+      estimatedNextRequestTokens: input.estimatedNextRequestTokens,
+      contextWindow: state.contextWindow,
+      reserveTokens,
+      reserveTailEvents: midTurn.reserveTailEvents ?? 1,
+      charsPerToken,
+      now: this.now(),
+      ...(compactPolicy.highWaterName !== undefined
+        ? { highWaterName: compactPolicy.highWaterName }
+        : {}),
+      ...(state.previousCheckpoint ? { previousCheckpoint: state.previousCheckpoint } : {}),
+      summarize: async ({ coveredRuntimeEvents, newlyFoldedRuntimeEvents, previousCheckpoint }) => {
+        return await Promise.resolve(
+          summarizer({
+            sessionId: this.sessionId,
+            turnId,
+            source: { foldedRuntimeEvents: [...coveredRuntimeEvents] },
+            ...(previousCheckpoint ? { previousCheckpoint } : {}),
+            newlyFoldedRuntimeEvents: [...newlyFoldedRuntimeEvents],
+            ...(abortSignal ? { abortSignal } : {}),
+          }),
+        );
+      },
+    });
+
+    if (plan.decision === 'skip') return { decision: 'skip' };
+    if (plan.decision === 'fail_open') {
+      return {
+        decision: 'fail',
+        detail: plan.reason,
+        diagnosticReason: plan.diagnosticReason ?? plan.reason,
+      };
+    }
+
+    // Lifecycle order is validate → persist → apply, where validate =
+    // materializable ∧ smaller ∧ replay-admissible. Replay applies the
+    // session's latest checkpoint BEFORE any high-water check, so a
+    // checkpoint that fails ANY of the three must never be persisted — it
+    // would poison every later projection even though this step correctly
+    // refused it.
+    const replayPlan = buildRuntimeEventModelReplayPlan(plan.replacementEvents, {
+      toolActivityTurnIds: collectToolActivityTurnIds(orderedEvents),
+    });
+    if (
+      replayPlan.items.length === 0 ||
+      hasBlockingReplayDiagnostics(replayPlan) ||
+      (replayPlan.hasProviderNativeSemantics && !this.canReplayProviderNative(replayPlan))
+    ) {
+      return {
+        decision: 'fail',
+        detail: 'no_safe_completed_span',
+        diagnosticReason: 'replacement_unmaterializable',
+      };
+    }
+    // The head anchor must render exactly like the raw projection's current
+    // user message: the initial request decorates it with the volatile turn
+    // tail (cwd, shell context, task state — see send()), which is not part
+    // of the durable anchor bytes. Reuse the same decoration owner
+    // (appendTurnTailPrompt) on the anchor's replay item so a replacement
+    // never silently drops that context — and never counts the drop as
+    // shrinkage in the guard below.
+    const replayItemsWithAnchorTail = replayPlan.items.map((item) =>
+      item.kind === 'text' && item.role === 'user' && item.eventId === state.headAnchor.id
+        ? { ...item, content: this.appendTurnTailPrompt(item.content, turnTailPrompt) as string }
+        : item,
+    );
+    const replacementMessages = await this.materializeRuntimeReplayPlan({
+      ...replayPlan,
+      items: replayItemsWithAnchorTail,
+    });
+    // Apply the shape only when it actually shrinks the request versus the
+    // reference payload (the incoming request for the proactive hook, the
+    // request that overflowed for reactive recovery): a materialized
+    // replacement that is not smaller proves the summarizer's OUTPUT is
+    // unusable, reported as summarizer_failed via replacement_not_smaller.
+    const replacedPayloadChars = midTurnRequestPayloadChars(
+      replacementMessages,
+      providerTools,
+      activeToolsForStep,
+      systemPromptChars,
+    );
+    if (replacedPayloadChars >= input.referencePayloadChars) {
+      return {
+        decision: 'fail',
+        detail: 'summarizer_failed',
+        diagnosticReason: 'replacement_not_smaller',
+      };
+    }
+    // Replay admissibility uses the same complete-prefix capacity gate as
+    // recovery. Actual payload shrinkage was already checked above because
+    // only this owner can measure the fully materialized provider request.
+    const replayFit = evaluateHistoryCompactCheckpointReplay(
+      plan.checkpoint,
+      plan.replacementEvents.slice(1),
+      policy?.charsPerToken,
+      policy?.maxHistoryEstimatedTokens,
+    );
+    if (!replayFit.fits) {
+      return {
+        decision: 'fail',
+        detail: 'head_anchor_exceeds_capacity',
+        diagnosticReason: `replay_rejected_${replayFit.reason}`,
+      };
+    }
+
+    // The replacement is valid: durably persist the checkpoint BEFORE
+    // applying the projection — the same order as the pre_turn path. A
+    // persistence failure keeps raw messages and records write_failed.
+    try {
+      await Promise.resolve(recorder(plan.checkpoint, turnId));
+    } catch {
+      return {
+        decision: 'fail',
+        detail: 'summarizer_failed',
+        diagnosticReason: 'write_failed',
+        recorderCounters: { historyCompactWritesAttempted: 1, historyCompactWriteFailures: 1 },
+      };
+    }
+    state.previousCheckpoint = plan.checkpoint;
+    return {
+      decision: 'compacted',
+      checkpoint: plan.checkpoint,
+      replacementMessages,
+      estimatedTokensBefore: plan.estimatedTokensBefore,
+      estimatedTokensAfter: plan.estimatedTokensAfter,
+    };
+  }
+
+  /**
+   * Reactive overflow recovery (issue #882 PR 2): the second line of defense.
+   * When a provider rejects a request with a context-length error, fold the
+   * durable turn ledger once and resend once — a single compact-and-retry
+   * latch (pi's `_overflowRecoveryAttempted`). Returns the compacted messages
+   * to resend, or undefined when recovery is impossible or already spent, in
+   * which case the caller surfaces the real provider error rather than a
+   * fabricated success or a synthesized `context_budget_exhausted` (the
+   * provider — not the runtime — rejected the request). Non-context-length
+   * errors and turns without the mid-turn seam never reach compaction, so the
+   * default (no seam) behavior is already better than the old fake end_turn.
+   */
+  public async recoverFromOverflowError(input: {
+    error: unknown;
+    retryAlreadyUsed: boolean;
+    midTurnState: MidTurnCapacityCompactState | undefined;
+    turnId: string;
+    currentMessages: readonly ModelMessage[];
+    providerTools: readonly MakaTool[];
+    activeTools: readonly string[];
+    systemPromptChars: number;
+    turnTailPrompt: string | undefined;
+    queue: AsyncEventQueue<SessionEvent>;
+    onDiagnosticPatch: (patch: Partial<ContextBudgetDiagnostic>) => void;
+    abortSignal?: AbortSignal;
+  }): Promise<{ messages: ModelMessage[] } | undefined> {
+    const state = input.midTurnState;
+    if (input.retryAlreadyUsed || !state) return undefined;
+    if (this.modelAdapter.classifyError(input.error) !== 'ContextLength') return undefined;
+
+    // The shrink baseline is the request the provider actually rejected. Its
+    // single owner is the verdict owner's per-request payload measure
+    // (state.lastRequestPayloadChars), recorded at the end of every
+    // prepareStep run — the attempt-INITIAL messages undercount the rejected
+    // request by every same-turn tool step, and a baseline anchored there
+    // refuses folds that genuinely shrink the real request (review P1-1).
+    // The cold-start fallback only covers a send whose verdict owner never
+    // ran a prepareStep (defensive; step 0 records the baseline too).
+    const referencePayloadChars =
+      state.lastRequestPayloadChars ??
+      midTurnRequestPayloadChars(
+        input.currentMessages,
+        input.providerTools,
+        input.activeTools,
+        input.systemPromptChars,
+      );
+    const outcome = await this.computeMidTurnCompactionReplacement({
+      turnId: input.turnId,
+      state,
+      queue: input.queue,
+      // The stream has ended, so every completed step is already flushed; wait
+      // only for the consumer to drain the durable ledger up to date.
+      minFlushedSteps: state.flushedSteps,
+      // The provider rejected the request outright, so force the fold past the
+      // high water regardless of the (evidently under-counting) estimate.
+      estimatedNextRequestTokens: state.contextWindow + 1,
+      referencePayloadChars,
+      providerTools: input.providerTools,
+      activeToolsForStep: input.activeTools,
+      systemPromptChars: input.systemPromptChars,
+      turnTailPrompt: input.turnTailPrompt,
+      abortSignal: input.abortSignal,
+    });
+    if (outcome.decision !== 'compacted') {
+      // Recovery attempted but could not produce a smaller, admissible
+      // request; record the failed overflow attempt and let the caller surface
+      // the real provider error.
+      input.onDiagnosticPatch({
+        historyCompactEnabled: true,
+        historyCompactMode: 'read_write',
+        ...(outcome.decision === 'fail' && outcome.recorderCounters
+          ? outcome.recorderCounters
+          : {}),
+        ...compactionDecisionDiagnosticPatch({
+          stage: 'activeStep',
+          sourceKind: 'runtimeEvents',
+          decision: 'failedOpen',
+          phase: 'mid_turn',
+          boundaryKind: 'historyCompact',
+          reason: 'overflow',
+          ...(outcome.decision === 'fail'
+            ? {
+                failOpenReason: outcome.diagnosticReason,
+                skippedReasonCounts: { [outcome.diagnosticReason]: 1 },
+              }
+            : {}),
+        }),
+      });
+      return undefined;
+    }
+    input.onDiagnosticPatch(
+      buildMidTurnReplacedDiagnosticPatch({
+        checkpoint: outcome.checkpoint,
+        estimatedTokensBefore: outcome.estimatedTokensBefore,
+        estimatedTokensAfter: outcome.estimatedTokensAfter,
+        reason: 'overflow',
+      }),
+    );
+    // A successful recovery restructures the request, so the rejected
+    // request's payload measure no longer describes what the retry sends.
+    // Reset the baseline: the capacity hook's usage anchor is only coherent
+    // paired with the payload chars of the SAME request, and a missing
+    // baseline forces the whole-payload cold-start estimate instead of a
+    // stale pairing against the dead attempt.
+    state.lastRequestPayloadChars = undefined;
+    return { messages: outcome.replacementMessages };
+  }
+
+  /**
+   * The single end-of-pipeline estimate owner for the mid-turn capacity
+   * invariant. Every prepareStep hook only shapes; this wrapper measures the
+   * FINAL outgoing (messages, tools) payload — the bytes the provider will
+   * actually see, after capacity compaction, active tool-result pruning, and
+   * semantic/active-full compaction have all run — and issues the one
+   * safety-critical verdict:
+   *
+   *  - estimate = the last request's real INPUT tokens + signed char/4 delta
+   *    against the previous request's measured payload (recorded here on
+   *    every step, including step 0's baseline); the delta already carries
+   *    the step's fresh output, so an output-inclusive baseline would count
+   *    it twice, and an unusable usage sample falls back to the whole-payload
+   *    cold start rather than a zero baseline;
+   *  - over the window with no capacity attempt this step (the approximate
+   *    trigger missed, e.g. growth the trigger under-weighted), force ONE
+   *    capacity re-entry — the verdict must not terminate a turn a shaper can
+   *    still rescue, and one bounded re-entry preserves termination;
+   *  - still over the window → context_budget_exhausted, with the terminal
+   *    detail taken from this step's capacity outcome: a replacement that
+   *    remains too large is head_anchor_exceeds_capacity (the irreducible
+   *    remainder exceeds capacity); a recorded shaping failure keeps its own
+   *    detail and diagnostic reason.
+   *
+   * Step 0 is shaped by the pre_turn path and only records the baseline here.
+   */
+  public buildMidTurnFinalRequestVerdict(input: {
+    shaped: PrepareStepFunctionLike;
+    reentry: PrepareStepFunctionLike;
+    state: MidTurnCapacityCompactState;
+    providerTools: readonly MakaTool[];
+    fallbackActiveTools: () => readonly string[];
+    charsPerToken: number;
+    systemPromptChars: number;
+    onDiagnosticPatch: (patch: Partial<ContextBudgetDiagnostic>) => void;
+    abortController?: AbortController | null;
+  }): PrepareStepFunctionLike {
+    const {
+      shaped,
+      reentry,
+      state,
+      providerTools,
+      fallbackActiveTools,
+      charsPerToken,
+      systemPromptChars,
+      onDiagnosticPatch,
+      abortController,
+    } = input;
+    return async (options) => {
+      let result = await Promise.resolve(shaped(options));
+      const finalPayloadChars = (): number =>
+        midTurnRequestPayloadChars(
+          result?.messages ?? options.messages,
+          providerTools,
+          result?.activeTools ?? options.activeTools ?? fallbackActiveTools(),
+          systemPromptChars,
+        );
+      let payloadChars = finalPayloadChars();
+      if (options.stepNumber >= 1 && !state.exhaustedDetail) {
+        const estimateFinal = (): number =>
+          estimateNextRequestTokens({
+            ...(state.lastRequestInputTokens !== undefined
+              ? { priorUsageTokens: state.lastRequestInputTokens }
+              : {}),
+            appendedChars: payloadChars - (state.lastRequestPayloadChars ?? payloadChars),
+            charsPerToken,
+            coldStartChars: payloadChars,
+          });
+        let estimate = estimateFinal();
+        const capacityAttemptedThisStep =
+          state.replacedStepNumber === options.stepNumber ||
+          state.lastShapeFailure?.stepNumber === options.stepNumber;
+        if (estimate > state.contextWindow && !capacityAttemptedThisStep) {
+          // One bounded capacity re-entry: the trigger threshold is
+          // approximate on purpose (recoverable), so a miss must become a
+          // rescue attempt before it can become a terminal verdict. Re-run
+          // only the capacity + prune shapers over the already-shaped
+          // projection; a second attempt after a same-step failure is
+          // pointless (the failure was not a trigger miss) and would double
+          // recorder counters and summarizer calls.
+          state.forcedTriggerEstimate = estimate;
+          const reshaped = await Promise.resolve(
+            reentry({
+              ...options,
+              messages: result?.messages ?? options.messages,
+              ...(result?.activeTools ? { activeTools: result.activeTools } : {}),
+            }),
+          );
+          state.forcedTriggerEstimate = undefined;
+          if (reshaped) {
+            result = {
+              ...(result ?? {}),
+              ...reshaped,
+              activeTools: reshaped.activeTools ?? result?.activeTools,
+            };
+          }
+          payloadChars = finalPayloadChars();
+          estimate = estimateFinal();
+        }
+        if (estimate > state.contextWindow) {
+          const failure =
+            state.lastShapeFailure?.stepNumber === options.stepNumber
+              ? state.lastShapeFailure
+              : undefined;
+          const replacedThisStep = state.replacedStepNumber === options.stepNumber;
+          const detail: ContextBudgetExhaustedDetail = replacedThisStep
+            ? 'head_anchor_exceeds_capacity'
+            : (failure?.detail ?? 'no_safe_completed_span');
+          const diagnosticReason = replacedThisStep
+            ? 'head_anchor_exceeds_capacity'
+            : (failure?.diagnosticReason ?? 'no_safe_completed_span');
+          state.exhaustedDetail = detail;
+          onDiagnosticPatch({
+            historyCompactEnabled: true,
+            historyCompactMode: 'read_write',
+            ...compactionDecisionDiagnosticPatch({
+              stage: 'activeStep',
+              sourceKind: 'runtimeEvents',
+              decision: 'unchanged',
+              phase: 'mid_turn',
+              boundaryKind: 'historyCompact',
+              reason: 'context_budget_exhausted',
+              skippedReasonCounts: { [diagnosticReason]: 1 },
+            }),
+          });
+          abortController?.abort(new Error(`mid-turn context budget exhausted: ${detail}`));
+          return result;
+        }
+      }
+      state.lastRequestPayloadChars = payloadChars;
+      return result;
+    };
+  }
 }
 
 // -- moved helpers (defined in ai-sdk-backend, used only by cache write) -------
@@ -1339,5 +2086,196 @@ export function hasActiveToolResultPruneDiagnosticPatch(
     (patch.activePrunedToolResults ?? 0) > 0 ||
     (patch.activeArchiveFailures ?? 0) > 0 ||
     (patch.activeEstimatedTokensSaved ?? 0) > 0
+  );
+}
+
+// -- moved helpers (mid-turn / replay) ----------------------------------------
+
+export class MidTurnCapacityCompactState {
+  /**
+   * Chars of the final (system prompt + messages + active tool schema)
+   * payload of the LAST prepared request, recorded by the final-request
+   * estimate owner at the end of every prepareStep pipeline run. All capacity estimates are signed
+   * deltas against this number, so they are anchored to the request the
+   * provider actually saw — a compacted projection, a pruned tail, or a
+   * same-turn tool-schema expansion all move the delta the same way.
+   */
+  lastRequestPayloadChars: number | undefined;
+  /**
+   * The last request's REAL input size: the inputTokens the provider reported
+   * for the last finished step. Never input+output — the signed payload delta
+   * already carries the step's freshly generated output (assistant text/tool
+   * calls) and its tool results, so an output-inclusive baseline would count
+   * them twice. Undefined when the last step's usage is missing or unusable
+   * (no positive input count); estimates then fall back to the whole-payload
+   * cold-start path — an unusable sample is unknown, never zero.
+   */
+  lastRequestInputTokens: number | undefined;
+  /** Latest durable checkpoint (loaded or written) for roll-forward summaries. */
+  previousCheckpoint: HistoryCompactCheckpoint | undefined;
+  /** Set when the turn must end with a context_budget_exhausted outcome. */
+  exhaustedDetail: ContextBudgetExhaustedDetail | undefined;
+  /**
+   * Step whose request the capacity hook replaced. Semantic/active-full
+   * compaction yields on that exact step so one step never runs two
+   * summarizers or double-projects.
+   */
+  replacedStepNumber: number | undefined;
+  /**
+   * finish-step boundaries the event pump has flushed into the session-event
+   * queue. The capacity hook's durability wait needs it: only after the pump
+   * has flushed step N's boundary are that step's thinking/text completion
+   * events enqueued at all.
+   */
+  flushedSteps = 0;
+  /**
+   * Set by the final-request estimate owner to force one capacity re-entry on
+   * the current step, bypassing the (deliberately approximate) high-water
+   * trigger. Consumed by the capacity hook on its next invocation.
+   */
+  forcedTriggerEstimate: number | undefined;
+  /**
+   * The capacity hook's most recent shaping failure. The owner reads it (for
+   * the same step only) to pick the terminal detail and diagnostic reason
+   * when the final payload is over the window, and to avoid re-entering a
+   * shaper that already attempted and failed this step.
+   */
+  lastShapeFailure:
+    | {
+        stepNumber: number;
+        detail: ContextBudgetExhaustedDetail;
+        diagnosticReason: string;
+      }
+    | undefined;
+
+  constructor(
+    readonly headAnchor: RuntimeEvent,
+    readonly priorContentEvents: readonly RuntimeEvent[],
+    readonly contextWindow: number,
+  ) {}
+}
+
+/**
+ * Char measure of the FULL provider-visible request input: the system prompt
+ * (sent through the separate `system` field), the (projected) messages, and
+ * the serialized schemas of the active tool subset. The capacity trigger and
+ * the final-request estimate owner both measure with this ONE function, so
+ * their deltas against `lastRequestPayloadChars` are commensurable and
+ * same-turn tool-schema growth (a `load_tools` activation) is counted like
+ * any other payload growth. The system prompt is constant between adjacent
+ * requests — signed deltas cancel it — but the cold-start estimate (no usable
+ * usage sample) is the whole payload, so omitting it would under-estimate by
+ * exactly the system prompt and let an over-window request stream.
+ */
+function midTurnRequestPayloadChars(
+  messages: readonly ModelMessage[],
+  providerTools: readonly MakaTool[],
+  activeTools: readonly string[],
+  systemPromptChars: number,
+): number {
+  return (
+    Math.max(0, Math.floor(systemPromptChars)) +
+    JSON.stringify(messages).length +
+    toolSchemaCharsForDiagnostics(providerTools, activeTools)
+  );
+}
+
+/**
+ * Outcome of folding the durable turn ledger into a replacement projection.
+ * Shared by the proactive prepareStep hook (which maps it to keepProjection /
+ * shapeFailure / a `context_limit` replacement) and the reactive overflow
+ * recovery (which maps it to a retry / a real error terminal, with an
+ * `overflow` reason). The verdict/diagnostic is the caller's; this only shapes.
+ */
+type MidTurnCompactionOutcome =
+  | { decision: 'skip' }
+  | {
+      decision: 'fail';
+      detail: ContextBudgetExhaustedDetail;
+      diagnosticReason: string;
+      recorderCounters?: Partial<ContextBudgetDiagnostic>;
+    }
+  | {
+      decision: 'compacted';
+      checkpoint: HistoryCompactCheckpoint;
+      replacementMessages: ModelMessage[];
+      estimatedTokensBefore: number;
+      estimatedTokensAfter: number;
+    };
+
+/**
+ * The `decision: 'replaced'` diagnostic patch for a durable mid_turn fold,
+ * shared by the proactive (`reason: 'context_limit'`) and reactive
+ * (`reason: 'overflow'`) triggers so both report the fold identically.
+ */
+function buildMidTurnReplacedDiagnosticPatch(input: {
+  checkpoint: HistoryCompactCheckpoint;
+  estimatedTokensBefore: number;
+  estimatedTokensAfter: number;
+  reason: string;
+}): Partial<ContextBudgetDiagnostic> {
+  const { checkpoint, estimatedTokensBefore, estimatedTokensAfter, reason } = input;
+  return {
+    historyCompactEnabled: true,
+    historyCompactMode: 'read_write',
+    historyCompactWritesAttempted: 1,
+    historyCompactBlocksWritten: 1,
+    historyCompactWrittenBlockIds: [checkpoint.checkpointId],
+    historyCompactWriteEstimatedTokens: checkpoint.estimatedTokens,
+    historyCompactBlockIds: [checkpoint.checkpointId],
+    historyCompactedTurns: checkpoint.coverage.turnCount,
+    historyCompactedEvents: checkpoint.coverage.eventCount,
+    historyCompactedEstimatedTokensBefore: estimatedTokensBefore,
+    historyCompactedEstimatedTokensAfter: estimatedTokensAfter,
+    highWaterName: checkpoint.highWaterName,
+    highWaterSeq: checkpoint.highWaterSeq,
+    highWaterReason: 'history_compact',
+    ...compactionDecisionDiagnosticPatch({
+      stage: 'activeStep',
+      sourceKind: 'runtimeEvents',
+      decision: 'replaced',
+      phase: 'mid_turn',
+      boundaryKind: 'historyCompact',
+      boundaryIds: [checkpoint.checkpointId],
+      coverage: { bodySha256: [checkpoint.coverage.sourceDigest] },
+      reason,
+      estimatedTokensBefore,
+      estimatedTokensAfter,
+    }),
+  };
+}
+
+/**
+ * Event-driven wait for seq-ack progress: resolves when the queue reports any
+ * push/ack/close/wake, or immediately on abort. The caller loops and re-checks
+ * its condition — a condition variable, not a poll.
+ */
+function waitForQueueProgressOrAbort(
+  queue: AsyncEventQueue<SessionEvent>,
+  abortSignal: AbortSignal | undefined,
+): Promise<void> {
+  return new Promise<void>((resolve) => {
+    let settled = false;
+    const settle = (): void => {
+      if (settled) return;
+      settled = true;
+      abortSignal?.removeEventListener('abort', settle);
+      resolve();
+    };
+    abortSignal?.addEventListener('abort', settle, { once: true });
+    void queue.waitForProgress().then(settle);
+  });
+}
+
+export function hasBlockingReplayDiagnostics(plan: RuntimeEventModelReplayPlan): boolean {
+  // `unmatched_tool_result` is deliberately NOT blocking: the materializer
+  // drops an orphan tool result (its call sliced away or the ledger corrupt)
+  // on its own — see pushToolResults — so one orphan must not degrade the
+  // whole ledger to stored-message projection.
+  return plan.diagnostics.some(
+    (diagnostic) =>
+      diagnostic.code === 'unsupported_role' ||
+      diagnostic.code === 'unsupported_content' ||
+      diagnostic.code === 'tool_id_mismatch',
   );
 }
